Two Iterations 1)Changed in class ProjectCreatedConsumer log for get consumer 2)Deleted from class ProfileService method createProject and created class ProjectSevice for create project 3)Changed name method mapper 4)Deleted sql files
