+++ src/main/java/com/itmentorcommunityplatform/profileservice/httpRequest/createProjectConsumerRequests.http
@@ -1,13 +1,13 @@
 POST http://localhost:8084//api/projects/project
 Content-Type: application/json
-X-Telegram-User-Id: 123123123
+X-Telegram-User-Id: 1212121212
 X-Telegram-Username: telegram_dev_user
 
 {
-  "author_telegram_user_id": 12321321,
-  "github_repository_url": "https://github.com/zhukovsd/cloude",
+  "author_telegram_user_id": 1212121212,
+  "github_repository_url": "https://github.com/zhukovsd/clou",
   "programming_language": "Java",
   "roadmap_project": "SIMULATION",
-  "telegram_user_id": 2222,
+  "telegram_user_id": 12144444,
   "added_timestamp:": 123313
 }
