+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -3,3 +3,5 @@ databaseChangeLog:
       file: db/changelog/changesets/001-create-tables.sql
   - include:
       file: db/changelog/changesets/002-create-table.sql
+  - include:
+      file: db/changelog/changesets/003-create-table-project.sql
